Weather Buddy - Sample FreshDesk App.
This sample app displays the weather for the support agent in the Freshdesk platform ticket page. The app also allows an agent to pick an alternate city and a future date (within 5 days) and display the weather for the same. 

The agent will only be interested in 2 weather parameters - Temperature and Wind Speed.